Create script that backs up an Explorer session
With "session", I am just referring to the total collection of all
folders currently opened in Windows Explorer.

This script takes in a folder path, gets all the currently opened
windows, prints them out, and then writes a new, time-stamped text file
to said path containing the paths to the opened folders, one in each
line.

This is already usable on its own, to see what we had open at particular
date-times (which we COULD then manually open again, if we wanted to).