Add script to restore an Explorer session
This script takes the path to the folder with the session backup text
files, and can also take the name of a specific session backup file,
and then re-opens all the folders listed in said file.

The backup file name is optional. If none is specified, the script
gathers all the backup files (only .txt files are considered), sorts
them, and then uses the most recent one.

This seems to work fine right now! Tested with up to 16 open folders.

What I don't like is that there is no output right now about whether the
operation of opening a folder succeeded.
Also, we had to hard-code a delay in between each folder opening
operation. Ideally, we would wait for each one to finish and then
continue afterward.
Furthermore, the folders are all getting spammed into the foreground. It
would be nice if they were instead opened in the background.