Improve Explorer session restoration process
We now actually wait for each Explorer window to fully open before we
continue, instead of having a hard-coded delay.

We now also launch a separate process for each Explorer window. This
does of course use more system memory, but it should improve robustness
in case of one of the windows crashing. (It is also possible that this
only works if the "start a new process for each window" option for
Explorer is set in Windows, and that if it isn't, Windows consolidates
the Explorer windows into fewer processes anyway. I tried testing it
briefly, but couldn't see a difference. It is possible that this option
is broken right now, anyway.)

Furthermore, we now minimize the windows automatically after opening
them. It is also possible to open them minimized right away, but this
comes with its own issue (no taskbar previews shown until a window is
brought to the foreground manually).

This solves most of the issues we still had with this process,
previously.